+++ src/org/helioviewer/jhv/events/JHVRelatedEvents.java
@@ -20,7 +20,7 @@ import org.helioviewer.jhv.timelines.draw.ClickableDrawable;
 
 public class JHVRelatedEvents implements ClickableDrawable {
 
-    private static final Colors.BrightColors eventColors = new Colors.BrightColors();
+    private static final Colors.DataColors eventColors = new Colors.DataColors();
     private static final ArrayList<JHVEventListener.Highlight> listeners = new ArrayList<>();
 
     private final ArrayList<JHVEvent> events = new ArrayList<>();